+++ middleware_key_expired_check.go
@@ -34,6 +34,16 @@ func (k *KeyExpired) ProcessRequest(w http.ResponseWriter, r *http.Request, conf
 			"key":    authHeaderValue,
 		}).Info("Attempted access from expired key.")
 
+		// Fire a key expired event
+		authHeaderValue := context.Get(r, AuthHeaderValue)
+		go k.TykMiddleware.FireEvent(EVENT_KeyExpired,
+			EVENT_KeyExpiredMeta{
+			EventMetaDefault: EventMetaDefault{Message: "Attempted access from expired key."},
+			Path: r.URL.Path,
+			Origin: r.RemoteAddr,
+			Key: authHeaderValue.(string),
+		})
+
 		return errors.New("Key has expired, please renew"), 403
 	}