+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - API Requests require an api_id form value (either param or body) as keys are now stored (federated) on a per API basis and can live in multiple stores. THIS IS A BREAKING CHANGE
 - Will not work with Dashboard 0.7, don't even try it... will be updated separately.
 - Added IP white listing middleware, completely self-contained, very rudimentary, only explicit IP's allowed (doesn't handle subnets etc.), hould work with IPv4 and IPv6
+- Added an Event subsystem and framework for extending and creating custom event handlers (e.g. for webhooks)